rag_system/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.12
langchain-community==0.0.27
chromadb==0.4.24
pydantic==2.6.3
python-multipart==0.0.9
tiktoken==0.6.0
sentence-transformers==2.5.1
timeout-decorator==0.5.0
tenacity==8.2.3 